Add option to avoid using output directory to skip processes.
When everything is present in the work directory, the resum option
will do a perfect job of not re-running processes that were already
run. Using files from the output directory to skip parts of the pipeline
and use them as inputs for other processes will break nextflow's
ability to check whether these processes were already run, as the
input comes from somewhere else. These alternative entry points
are therefore nice when we do not have the work directory, but they
should not mess with nextflow's resume option, hence the option
to deactivate it.