Add script to copy input files from a source to the input folder.
This script is used to copy necessary input files from a source folder to the
pipeline input folder. It will only copy over input files for samples that have
not been analysed yet. It also tries to avoid copying over files that are currently
being downloaded in the source folder.